[2016.11.11] - Refactor;  - Renaming schemas and models accordingly;
